central_monitor_datamatrix/requirements.txt
numpy
opencv-python
pillow
pylibdmtx
